Carpeta adminid en las vistas

+++ routes/web.php
@@ -46,29 +46,29 @@ Route::get('/usuarios/usuariorestore/{usuario}', 'UserController@usuarioRestore'
 
 
 // Productos
-Route::get('/productos/index', 'ProductController@index')->name('productos.index')->middleware('auth');
-Route::get('/productos/create', 'ProductController@create')->name('productos.create')->middleware('auth');
-Route::post('/productos/store', 'ProductController@store')->name('productos.store')->middleware('auth');
-Route::get('/productos/show/{producto}', 'ProductController@show')->name('productos.show')->middleware('auth');
-Route::get('/productos/edit/{producto}', 'ProductController@edit')->name('productos.edit')->middleware('auth');
-Route::put('/productos/update/{producto}', 'ProductController@update')->name('productos.update')->middleware('auth');
-Route::delete('/productos/delete/{producto}', 'ProductController@delete')->name('productos.delete')->middleware('auth');
-Route::get('/productos/productoeliminado', 'ProductController@indexDelete')->name('productos.indexDelete')->middleware('auth');
-Route::get('/productos/productorestore/{producto}', 'ProductController@productoRestore')->name('productos.productorestore')->middleware('auth');
+Route::get('/productos/index', 'ProductController@index')->name('admin.productos.index')->middleware('auth');
+Route::get('/productos/create', 'ProductController@create')->name('admin.productos.create')->middleware('auth');
+Route::post('/productos/store', 'ProductController@store')->name('admin.productos.store')->middleware('auth');
+Route::get('/productos/show/{producto}', 'ProductController@show')->name('admin.productos.show')->middleware('auth');
+Route::get('/productos/edit/{producto}', 'ProductController@edit')->name('admin.productos.edit')->middleware('auth');
+Route::put('/productos/update/{producto}', 'ProductController@update')->name('admin.productos.update')->middleware('auth');
+Route::delete('/productos/delete/{producto}', 'ProductController@delete')->name('admin.productos.delete')->middleware('auth');
+Route::get('/productos/productoeliminado', 'ProductController@indexDelete')->name('admin.productos.indexDelete')->middleware('auth');
+Route::get('/productos/productorestore/{producto}', 'ProductController@productoRestore')->name('admin.productos.productorestore')->middleware('auth');
 
 //Galeria Productos
 // Route::post('productosgaleria/shore/{productogaleria}', 'ProductController@productGalery')->name('productosgaleria.store')->middleware('auth');
 
 
 // Categorias
-Route::get('/categorias/index', 'CategoryController@index')->name('categorias.index')->middleware('auth');
-Route::get('/categorias/create', 'CategoryController@create')->name('categorias.create')->middleware('auth');
-Route::post('/categorias/store', 'CategoryController@store')->name('categorias.store')->middleware('auth');
-Route::get('/categorias/edit/{categoria}', 'CategoryController@edit')->name('categorias.edit')->middleware('auth');
-Route::put('/categorias/update/{categoria}', 'CategoryController@update')->name('categorias.update')->middleware('auth');
-Route::delete('/categorias/delete/{categoria}', 'CategoryController@delete')->name('categorias.delete')->middleware('auth');
-Route::get('/categorias/indexdelete', 'CategoryController@indexDelete')->name('categorias.indexdelete')->middleware('auth');
-Route::get('/categorias/categoriarestore/{categoria}', 'CategoryController@categoriaRestore')->name('categorias.categoriarestore')->middleware('auth');
+Route::get('/categorias/index', 'CategoryController@index')->name('admin.categorias.index')->middleware('auth');
+Route::get('/categorias/create', 'CategoryController@create')->name('admin.categorias.create')->middleware('auth');
+Route::post('/categorias/store', 'CategoryController@store')->name('admin.categorias.store')->middleware('auth');
+Route::get('/categorias/edit/{categoria}', 'CategoryController@edit')->name('admin.categorias.edit')->middleware('auth');
+Route::put('/categorias/update/{categoria}', 'CategoryController@update')->name('admin.categorias.update')->middleware('auth');
+Route::delete('/categorias/delete/{categoria}', 'CategoryController@delete')->name('admin.categorias.delete')->middleware('auth');
+Route::get('/categorias/indexdelete', 'CategoryController@indexDelete')->name('admin.categorias.indexdelete')->middleware('auth');
+Route::get('/categorias/categoriarestore/{categoria}', 'CategoryController@categoriaRestore')->name('admin.categorias.categoriarestore')->middleware('auth');
 
 });
